scikit/requirements.txt
scikit-learn
matplotlib
pandas
mecab-python-windows
# janome
gensim

seaborn
